back/requirements.txt
Django>=3.0,<4.0
djangorestframework
pygments
psycopg2>=2.8
django-cors-headers
